Removed test... identified the invisible bug.
The "bug" in Myo Connect doesn't exist. I conclude that it was caused by
an issue in old myo_mex.cpp code related to DataCollector in global
scope and "clear myo_mex" not being called in MATLAB. This results in
DataCollector remaining in memory after "myo_mex delete" and stale Myo*
hanging on DataCollector.