Frame based streaming works with one Myo
Fixed optimization bug in configuration parameters of the example SLX
model. With optimizations and such turned back on, the model runs for
reasonable durations without the buffers draining. This was likely the
culprit of issues due to inefficiency of execution.

At this point, users should tune the value of BUFFER_FRAMES_DES to suit
their use. The default, 25, initializes te data buffers with 25 frames
(at 25Hz) resulting in 1s latency. Reducing this value results in
reduced latency (=0.04*BUFFER_FRAMES_DES) with increased risk of errors
due to draining the buffers.